// src/@v2/documents/domain/enums/document-children-type.enum.ts
export enum DocumentChildrenTypeEnum {
  TITLE = 'TITLE',
  H1 = 'H1',
  H2 = 'H2',
  H3 = 'H3',
  H4 = 'H4',
  H5 = 'H5',
  H6 = 'H6',
  PARAGRAPH = 'PARAGRAPH',
  IMAGE = 'IMAGE',
  BREAK = 'BREAK',
  BULLET = 'BULLET',
  BULLET_SPACE = 'BULLET_SPACE',
  TABLE_VERSION_CONTROL = 'TABLE_VERSION_CONTROL',
  ATTACHMENTS = 'ATTACHMENTS',
  PARAGRAPH_TABLE = 'PARAGRAPH_TABLE',
  LEGEND = 'LEGEND',
  PROFESSIONAL = 'PROFESSIONAL',
  PROFESSIONALS_SIGNATURES = 'PROFESSIONALS_SIGNATURES',
  PARAGRAPH_FIGURE = 'PARAGRAPH_FIGURE',

  ITERABLE_ENVIRONMENTS_ADM = 'ITERABLE_ENVIRONMENTS',
  ITERABLE_ENVIRONMENTS_OP = 'ITERABLE_ENVIRONMENTS_OP',
  ITERABLE_ENVIRONMENTS_SUP = 'ITERABLE_ENVIRONMENTS_SUP',
  ITERABLE_ENVIRONMENTS_GENERAL = 'ITERABLE_ENVIRONMENTS_GENERAL',
  ITERABLE_CHARACTERIZATION_EQUIP = 'ITERABLE_CHARACTERIZATION_EQUIP',
  ITERABLE_CHARACTERIZATION_ACTIVIT = 'ITERABLE_CHARACTERIZATION_ACTIVIT',
  ITERABLE_CHARACTERIZATION_WORKSTATION = 'ITERABLE_CHARACTERIZATION_WORKSTATION',
  ITERABLE_RECOMMENDATIONS = 'ITERABLE_RECOMMENDATIONS',
  ITERABLE_CONTROL_MEASURES = 'ITERABLE_CONTROL_MEASURES',
  ITERABLE_OTHER_CONTROL_MEASURES = 'ITERABLE_OTHER_CONTROL_MEASURES',
  ITERABLE_EPI_CONTROL_MEASURES = 'ITERABLE_EPI_CONTROL_MEASURES',
  ITERABLE_EMERGENCY_RISKS = 'ITERABLE_EMERGENCY_RISKS',
  TABLE_GSE = 'TABLE_GSE',
  TABLE_HIERARCHY_ENV = 'TABLE_HIERARCHY_ENV',
  TABLE_HIERARCHY_CHAR = 'TABLE_HIERARCHY_CHAR',
  TABLE_PRIORITIZATION = 'TABLE_PRIORITIZATION',
  TABLE_PRIORITIZATION_ENV = 'TABLE_PRIORITIZATION_ENV',
  TABLE_PRIORITIZATION_CHAR = 'TABLE_PRIORITIZATION_CHAR',
  TABLE_PRIORITIZATION_HIERARCHY = 'TABLE_PRIORITIZATION_HIERARCHY',

  ITERABLE_QUALITY_FIS = 'ITERABLE_QUALITY_FIS',
  ITERABLE_QUALITY_QUI = 'ITERABLE_QUALITY_QUI',
  ITERABLE_QUALITY_BIO = 'ITERABLE_QUALITY_BIO',
  ITERABLE_QUALITY_ERG = 'ITERABLE_QUALITY_ERG',
  ITERABLE_QUALITY_ACI = 'ITERABLE_QUALITY_ACI',

  TABLE_QUANTITY_NOISE = 'TABLE_QUANTITY_NOISE',
  TABLE_QUANTITY_QUI = 'TABLE_QUANTITY_QUI',
  TABLE_QUANTITY_HEAT = 'TABLE_QUANTITY_HEAT',
  TABLE_QUANTITY_VFB = 'TABLE_QUANTITY_VFB',
  TABLE_QUANTITY_VL = 'TABLE_QUANTITY_VL',
  TABLE_QUANTITY_RAD = 'TABLE_QUANTITY_RAD',

  HIERARCHY_RISK_TABLE = 'HIERARCHY_RISK_TABLE',
  RISK_TABLE = 'RISK_TABLE',
  MEASURE_IMAGE = 'MEASURE_IMAGE',
  RS_IMAGE = 'RS_IMAGE',
  COMPLEMENTARY_DOCS = 'COMPLEMENTARY_DOCS',
  COMPLEMENTARY_SYSTEMS = 'COMPLEMENTARY_SYSTEMS',
  HEALTH_EFFECT_TABLES = 'HEALTH_EFFECT_TABLES',
  EXPOSITION_DEGREE_TABLES = 'EXPOSITION_DEGREE_TABLES',
  HIERARCHY_ORG_TABLE = 'HIERARCHY_ORG_TABLE',
  QUANTITY_RESULTS_TABLES = 'QUANTITY_RESULTS_TABLES',
  QUANTITY_CONSIDERATION_TABLES = 'QUANTITY_CONSIDERATION_TABLES',
  MATRIX_TABLES = 'MATRIX_TABLES',
  APR_TABLE = 'APR_TABLE',
  PLAN_TABLE = 'PLAN_TABLE',

  TABLE_PCMSO_GHO = 'TABLE_PCMSO_GHO',
  TABLE_PCMSO_HIERARCHY = 'TABLE_PCMSO_HIERARCHY',
  TABLE_PCMSO_HIERARCHY_CONCAT = 'TABLE_PCMSO_HIERARCHY_CONCAT',

  PERICULOSIDADE_ACTIVITIES = 'PERICULOSIDADE_ACTIVITIES',
}